Finishing PayrollServiceImpl and add schema.sql to auto create table

+++ Project/Code/Backend/hr-management/src/main/java/com/hr_software_project/hr_management/dto/CreateUserRequestDTO.java
@@ -17,7 +17,7 @@ public class CreateUserRequestDTO {
     private String role;
     private Double salary;
     private Double daily_working_hours;
-    private Double number_of_working_days_per_week;
+    private Integer number_of_working_days_per_week;
     private String epfNumber;
     private String socsoNumber;
     private String incomeTaxNumber;
